flushed out the mysql plugin for get events.  sreport will have new reports for it now, and we should also update the documentation of the plugin to include the new function

+++ src/plugins/accounting_storage/mysql/mysql_jobacct_process.c
@@ -771,7 +771,7 @@ extern int setup_job_cond_limits(mysql_conn_t *mysql_conn,
 		while((object = list_next(itr))) {
 			if(set)
 				xstrcat(*extra, " || ");
-			_state_time_string(extra, atoi(object),
+			_state_time_string(extra, (uint32_t)atoi(object),
 					   job_cond->usage_start,
 					   job_cond->usage_end);
 			set = 1;
